refactor(python): use doubles uniformly

+++ inc/example/functions.h
@@ -6,8 +6,8 @@
 namespace py = pybind11;
 
 namespace example::functions {
-int add(int i, int j);
-int sub(int i, int j);
+double add(double i, double j);
+double sub(double i, double j);
 double mul(double i, double j);
 double div(double i, double j);
 }  // namespace example::functions
